+++ CHANGELOG.md
@@ -3,17 +3,23 @@
 ## [Unreleased]
 
 ### Added
-- Raise error if range gate indexing is inconsitent
-- Add plotting for beta and screened variables
 
 ### Changed
-- Screen values where intensity is below one
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+## [0.1.2] - 2023-04-28
+
+### Added
+- Raise error if range gate indexing is inconsitent
+- Add plotting for beta and screened variables
+
+### Changed
+- Screen values where intensity is below one
 
 ## [0.1.1] - 2023-04-25
 
@@ -78,3 +84,4 @@
 [0.0.9]: https://github.com/actris-cloudnet/halo-reader/releases/tag/v0.0.9
 [0.1.0]: https://github.com/actris-cloudnet/halo-reader/releases/tag/v0.1.0
 [0.1.1]: https://github.com/actris-cloudnet/halo-reader/releases/tag/v0.1.1
+[0.1.2]: https://github.com/actris-cloudnet/halo-reader/releases/tag/v0.1.2